+++ Sources/Tokenizer/Layout/Constraints/ConstraintType.swift
@@ -8,8 +8,8 @@
 import Foundation
 
 public enum ConstraintType {
-    case constant(Float)
-    case targeted(target: ConstraintTarget, targetAnchor: LayoutAnchor, multiplier: Float, constant: Float)
+    case constant(Double)
+    case targeted(target: ConstraintTarget, targetAnchor: LayoutAnchor, multiplier: Double, constant: Double)
 }
 
 extension ConstraintType: Equatable {
