+++ Sources/Tokenizer/Properties/Types/Implementation/Transformations/TransformationModifier.swift
@@ -9,9 +9,9 @@
 // FIXME Shouldn't this be XMLDeserializable/Serializable?
 public enum TransformationModifier {
     case identity
-    case rotate(by: Float)
-    case scale(byX: Float, byY: Float)
-    case translate(byX: Float, byY: Float)
+    case rotate(by: Double)
+    case scale(byX: Double, byY: Double)
+    case translate(byX: Double, byY: Double)
 }
 
 extension TransformationModifier {
